# processor/internal/gateway/docs/swagger.yaml
basePath: /
definitions:
  handler.APIError:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  pb.Account:
    properties:
      account_id:
        description: Unique identifier for the trading account.
        type: string
      account_type:
        description: Type of account (e.g., Stock, Future).
        type: string
      broker_id:
        description: Identifier for the broker.
        type: string
      person_id:
        description: Unique identifier for the account holder.
        type: string
      signed:
        description: Whether the account has signed the trading agreement.
        type: boolean
      username:
        description: User's name associated with the account.
        type: string
    type: object
...
  pb.Contract:
    properties:
      category:
        description: Sector or category.
        type: string
      code:
        description: Security code (e.g., stock number).
        type: string
      currency:
        allOf:
        - $ref: '#/definitions/pb.Currency'
        description: Currency used for trading.
      day_trade:
        allOf:
        - $ref: '#/definitions/pb.DayTrade'
        description: Day trade eligibility.
      delivery_date:
        description: Specific delivery date.
        type: string
      delivery_month:
        description: Delivery month (for derivatives).
        type: string
      exchange:
        allOf:
        - $ref: '#/definitions/pb.Exchange'
        description: Exchange where the security is traded.
      limit_down:
        description: Floor price for the day.
        type: number
      limit_up:
        description: Ceiling price for the day.
        type: number
      margin_trading_balance:
        description: Margin trading limit/balance.
        type: integer
      multiplier:
        description: Price multiplier.
        type: integer
      name:
        description: Full name of the security.
        type: string
      option_right:
        allOf:
        - $ref: '#/definitions/pb.OptionRight'
        description: Call or Put.
      reference:
        description: Yesterday's closing or reference price.
        type: number
      security_type:
        allOf:
        - $ref: '#/definitions/pb.SecurityType'
        description: Commodity category (Stock, Future, etc.).
      short_selling_balance:
        description: Short selling limit/balance.
        type: integer
      strike_price:
        description: Strike price (for options).
        type: number
      symbol:
        description: Ticker symbol.
        type: string
      target_code:
        description: Target underlying code.
        type: string
      underlying_code:
        description: Code of underlying asset.
        type: string
      underlying_kind:
        description: Type of underlying asset.
        type: string
      unit:
        description: Trading unit.
        type: number
      update_date:
        description: Date of last update.
        type: string
    type: object
  pb.CreditEnquire:
    properties:
      margin_unit:
        description: Available margin units.
        type: integer
      short_unit:
        description: Available short selling units.
        type: integer
      stock_id:
        description: Security code.
        type: string
      system:
        description: System code.
        type: string
      update_time:
        description: Update timestamp.
        type: string
    type: object
  pb.CreditEnquiresRequest:
    properties:
      contract_codes:
        items:
          type: string
        type: array
    type: object
  pb.CreditEnquiresResponse:
    properties:
      credit_enquires:
        items:
          $ref: '#/definitions/pb.CreditEnquire'
        type: array
    type: object
  pb.Currency:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    - 7
    - 8
    - 9
    - 10
    - 11
    - 12
    - 13
    - 14
    - 15
    - 16
    - 17
    - 18
    - 19
    - 20
    format: int32
    type: integer
    x-enum-comments:
      Currency_CURRENCY_AUD: Australian Dollar. 澳幣
      Currency_CURRENCY_CAD: Canadian Dollar. 加幣
      Currency_CURRENCY_CHF: Swiss Franc. 瑞士法郎
      Currency_CURRENCY_CNY: Chinese Yuan. 人民幣
      Currency_CURRENCY_EUR: Euro. 歐元
      Currency_CURRENCY_GBP: British Pound. 英鎊
      Currency_CURRENCY_HKD: Hong Kong Dollar. 港幣
      Currency_CURRENCY_IDR: Indonesian Rupiah. 印尼盾
      Currency_CURRENCY_JPY: Japanese Yen. 日圓
      Currency_CURRENCY_KRW: South Korean Won. 韓圓
      Currency_CURRENCY_MYR: Malaysian Ringgit. 馬來西亞幣
      Currency_CURRENCY_NZD: New Zealand Dollar. 紐西蘭幣
      Currency_CURRENCY_PHP: Philippine Peso. 菲律賓披索
      Currency_CURRENCY_SEK: Swedish Krona. 瑞典克朗
      Currency_CURRENCY_SGD: Singapore Dollar. 新加坡幣
      Currency_CURRENCY_THB: Thai Baht. 泰銖
      Currency_CURRENCY_TWD: New Taiwan Dollar. 新台幣
      Currency_CURRENCY_USD: US Dollar. 美元
      Currency_CURRENCY_VND: Vietnamese Dong. 越南盾
      Currency_CURRENCY_ZAR: South African Rand. 南非幣
    x-enum-descriptions:
    - ""
    - New Taiwan Dollar. 新台幣
    - US Dollar. 美元
    - Hong Kong Dollar. 港幣
    - British Pound. 英鎊
    - Australian Dollar. 澳幣
    - Canadian Dollar. 加幣
    - Singapore Dollar. 新加坡幣
    - Swiss Franc. 瑞士法郎
    - Japanese Yen. 日圓
    - South African Rand. 南非幣
    - Swedish Krona. 瑞典克朗
    - New Zealand Dollar. 紐西蘭幣
    - Thai Baht. 泰銖
    - Philippine Peso. 菲律賓披索
    - Indonesian Rupiah. 印尼盾
    - Euro. 歐元
    - South Korean Won. 韓圓
    - Vietnamese Dong. 越南盾
    - Malaysian Ringgit. 馬來西亞幣
    - Chinese Yuan. 人民幣
    x-enum-varnames:
    - Currency_CURRENCY_UNSPECIFIED
    - Currency_CURRENCY_TWD
    - Currency_CURRENCY_USD
    - Currency_CURRENCY_HKD
    - Currency_CURRENCY_GBP
    - Currency_CURRENCY_AUD
    - Currency_CURRENCY_CAD
    - Currency_CURRENCY_SGD
    - Currency_CURRENCY_CHF
    - Currency_CURRENCY_JPY
    - Currency_CURRENCY_ZAR
    - Currency_CURRENCY_SEK
    - Currency_CURRENCY_NZD
    - Currency_CURRENCY_THB
    - Currency_CURRENCY_PHP
    - Currency_CURRENCY_IDR
    - Currency_CURRENCY_EUR
    - Currency_CURRENCY_KRW
    - Currency_CURRENCY_VND
    - Currency_CURRENCY_MYR
    - Currency_CURRENCY_CNY
...
  pb.DayTrade:
    enum:
    - 0
    - 1
    - 2
    - 3
    format: int32
    type: integer
    x-enum-comments:
      DayTrade_DAY_TRADE_NO: No. 否
      DayTrade_DAY_TRADE_ONLYBUY: Buy first only. 先買後賣
      DayTrade_DAY_TRADE_YES: Yes. 可現沖
    x-enum-descriptions:
    - ""
    - Yes. 可現沖
    - Buy first only. 先買後賣
    - No. 否
    x-enum-varnames:
    - DayTrade_DAY_TRADE_UNSPECIFIED
    - DayTrade_DAY_TRADE_YES
    - DayTrade_DAY_TRADE_ONLYBUY
    - DayTrade_DAY_TRADE_NO
...
  pb.EarmarkStocksDetailResponse:
    properties:
      response_json:
        type: string
    type: object
...
  pb.Exchange:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    format: int32
    type: integer
    x-enum-comments:
      Exchange_EXCHANGE_OES: Emerging Stock. 興櫃
      Exchange_EXCHANGE_OTC: Taipei Exchange (Over-the-Counter). 上櫃
      Exchange_EXCHANGE_TAIFEX: Taiwan Futures Exchange. 期貨
      Exchange_EXCHANGE_TSE: Taiwan Stock Exchange. 上市
    x-enum-descriptions:
    - ""
    - Taiwan Stock Exchange. 上市
    - Taipei Exchange (Over-the-Counter). 上櫃
    - Emerging Stock. 興櫃
    - Taiwan Futures Exchange. 期貨
    x-enum-varnames:
    - Exchange_EXCHANGE_UNSPECIFIED
    - Exchange_EXCHANGE_TSE
    - Exchange_EXCHANGE_OTC
    - Exchange_EXCHANGE_OES
    - Exchange_EXCHANGE_TAIFEX
...
  pb.GetEarmarkingDetailRequest:
    properties:
      account:
        $ref: '#/definitions/pb.Account'
    type: object
...
  pb.GetShortStockSourcesRequest:
    properties:
      contract_codes:
        items:
          type: string
        type: array
    type: object
  pb.GetShortStockSourcesResponse:
    properties:
      sources:
        items:
          $ref: '#/definitions/pb.ShortStockSource'
        type: array
    type: object
...
  pb.GetStockReserveDetailRequest:
    properties:
      account:
        $ref: '#/definitions/pb.Account'
    type: object
  pb.GetStockReserveSummaryRequest:
    properties:
      account:
        $ref: '#/definitions/pb.Account'
    type: object
...
  pb.ListPositionDetailRequest:
    properties:
      account:
        $ref: '#/definitions/pb.Account'
      detail_id:
        type: integer
    type: object
  pb.ListPositionDetailResponse:
    properties:
      details:
        items:
          $ref: '#/definitions/pb.PositionDetail'
        type: array
    type: object
  pb.ListPositionsRequest:
    properties:
      account:
        $ref: '#/definitions/pb.Account'
      unit:
        description: Units (e.g., Common).
        type: string
    type: object
  pb.ListPositionsResponse:
    properties:
      positions:
        items:
          $ref: '#/definitions/pb.Position'
        type: array
    type: object
  pb.ListProfitLossDetailRequest:
    properties:
      account:
        $ref: '#/definitions/pb.Account'
      detail_id:
        type: integer
    type: object
  pb.ListProfitLossDetailResponse:
    properties:
      details:
        items:
          $ref: '#/definitions/pb.ProfitDetail'
        type: array
    type: object
  pb.ListProfitLossRequest:
    properties:
      account:
        $ref: '#/definitions/pb.Account'
      begin_date:
        description: Start date YYYYMMDD.
        type: string
      end_date:
        description: End date YYYYMMDD.
        type: string
    type: object
  pb.ListProfitLossResponse:
    properties:
      profit_losses:
        items:
          $ref: '#/definitions/pb.ProfitLoss'
        type: array
    type: object
  pb.ListProfitLossSummaryRequest:
    properties:
      account:
        $ref: '#/definitions/pb.Account'
    type: object
  pb.ListProfitLossSummaryResponse:
    properties:
      summaries:
        items:
          $ref: '#/definitions/pb.ProfitLossSummary'
        type: array
    type: object
...
  pb.OptionRight:
    enum:
    - 0
    - 1
    - 2
    - 3
    format: int32
    type: integer
    x-enum-comments:
      OptionRight_OPTION_RIGHT_CALL: Call option. 買權
      OptionRight_OPTION_RIGHT_NO: None. 無
      OptionRight_OPTION_RIGHT_PUT: Put option. 賣權
    x-enum-descriptions:
    - ""
    - None. 無
    - Call option. 買權
    - Put option. 賣權
    x-enum-varnames:
    - OptionRight_OPTION_RIGHT_UNSPECIFIED
    - OptionRight_OPTION_RIGHT_NO
    - OptionRight_OPTION_RIGHT_CALL
    - OptionRight_OPTION_RIGHT_PUT
...
  pb.Position:
    properties:
      position:
        description: "Types that are valid to be assigned to Position:\n\n\t*Position_StockPosition\n\t*Position_FuturePosition"
    type: object
  pb.PositionDetail:
    properties:
      detail:
        description: "Types that are valid to be assigned to Detail:\n\n\t*PositionDetail_StockDetail\n\t*PositionDetail_FutureDetail"
    type: object
  pb.ProfitDetail:
    properties:
      detail:
        description: "Types that are valid to be assigned to Detail:\n\n\t*ProfitDetail_StockDetail\n\t*ProfitDetail_FutureDetail"
    type: object
  pb.ProfitLoss:
    properties:
      item:
        description: "Types that are valid to be assigned to Item:\n\n\t*ProfitLoss_StockPnl\n\t*ProfitLoss_FuturePnl"
    type: object
  pb.ProfitLossSummary:
    properties:
      summary:
        description: "Types that are valid to be assigned to Summary:\n\n\t*ProfitLossSummary_StockSummary\n\t*ProfitLossSummary_FutureSummary"
    type: object
...
  pb.ReserveEarmarkingRequest:
    properties:
      account:
        $ref: '#/definitions/pb.Account'
      contract:
        $ref: '#/definitions/pb.Contract'
      price:
        type: number
      share:
        type: integer
    type: object
  pb.ReserveEarmarkingResponse:
    properties:
      response_json:
        type: string
    type: object
  pb.ReserveStockRequest:
    properties:
      account:
        $ref: '#/definitions/pb.Account'
      contract:
        $ref: '#/definitions/pb.Contract'
      share:
        type: integer
    type: object
  pb.ReserveStockResponse:
    properties:
      response_json:
        type: string
    type: object
  pb.ReserveStocksDetailResponse:
    properties:
      response_json:
        type: string
    type: object
  pb.ReserveStocksSummaryResponse:
    properties:
      response_json:
        type: string
    type: object
...
  pb.SecurityType:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    format: int32
    type: integer
    x-enum-comments:
      SecurityType_SECURITY_TYPE_FUT: Future. 期貨
      SecurityType_SECURITY_TYPE_IND: Index. 指數
      SecurityType_SECURITY_TYPE_OPT: Option. 選擇權
      SecurityType_SECURITY_TYPE_STK: Stock. 股票
    x-enum-descriptions:
    - ""
    - Index. 指數
    - Stock. 股票
    - Future. 期貨
    - Option. 選擇權
    x-enum-varnames:
    - SecurityType_SECURITY_TYPE_UNSPECIFIED
    - SecurityType_SECURITY_TYPE_IND
    - SecurityType_SECURITY_TYPE_STK
    - SecurityType_SECURITY_TYPE_FUT
    - SecurityType_SECURITY_TYPE_OPT
...
  pb.ShortStockSource:
    properties:
      code:
        description: Security code.
        type: string
      short_stock_source:
        description: Available shorting source volume.
        type: integer
      ts:
        description: Timestamp.
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API Gateway for Shioaji Trading Provider.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Phoenix API
  version: "1.0"
paths:
...
  /api/v1/positions:
    post:
      consumes:
      - application/json
      description: List all positions
      parameters:
      - description: Positions Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pb.ListPositionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.ListPositionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: List Positions
      tags:
      - Position
  /api/v1/positions/detail:
    post:
      consumes:
      - application/json
      description: Get detailed position information
      parameters:
      - description: Position Detail Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pb.ListPositionDetailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.ListPositionDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: List Position Detail
      tags:
      - Position
  /api/v1/positions/pnl:
    post:
      consumes:
      - application/json
      description: Get profit and loss information
      parameters:
      - description: P/L Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pb.ListProfitLossRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.ListProfitLossResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: List Profit Loss
      tags:
      - Position
  /api/v1/positions/pnl/detail:
    post:
      consumes:
      - application/json
      description: Get detailed profit and loss information
      parameters:
      - description: P/L Detail Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pb.ListProfitLossDetailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.ListProfitLossDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: List Profit Loss Detail
      tags:
      - Position
  /api/v1/positions/pnl/summary:
    post:
      consumes:
      - application/json
      description: Get profit and loss summary
      parameters:
      - description: P/L Summary Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pb.ListProfitLossSummaryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.ListProfitLossSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: List Profit Loss Summary
      tags:
      - Position
  /api/v1/reserve/credit-enquires:
    post:
      consumes:
      - application/json
      description: Enquire about credit
      parameters:
      - description: Credit Enquiries Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pb.CreditEnquiresRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.CreditEnquiresResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Credit Enquiries
      tags:
      - Reserve
  /api/v1/reserve/earmarking:
    post:
      consumes:
      - application/json
      description: Reserve funds for earmarking
      parameters:
      - description: Reserve Earmarking Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pb.ReserveEarmarkingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.ReserveEarmarkingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Reserve Earmarking
      tags:
      - Reserve
  /api/v1/reserve/earmarking/detail:
    post:
      consumes:
      - application/json
      description: Get earmarking details
      parameters:
      - description: Earmarking Detail Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pb.GetEarmarkingDetailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.EarmarkStocksDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Get Earmarking Detail
      tags:
      - Reserve
  /api/v1/reserve/short-stock-sources:
    post:
      consumes:
      - application/json
      description: Get sources for short selling
      parameters:
      - description: Short Stock Sources Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pb.GetShortStockSourcesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.GetShortStockSourcesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Get Short Stock Sources
      tags:
      - Reserve
  /api/v1/reserve/stocks:
    post:
      consumes:
      - application/json
      description: Reserve stock for trading
      parameters:
      - description: Reserve Stock Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pb.ReserveStockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.ReserveStockResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Reserve Stock
      tags:
      - Reserve
  /api/v1/reserve/stocks/detail:
    post:
      consumes:
      - application/json
      description: Get detailed stock reserve info
      parameters:
      - description: Reserve Detail Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pb.GetStockReserveDetailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.ReserveStocksDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Get Stock Reserve Detail
      tags:
      - Reserve
  /api/v1/reserve/stocks/summary:
    post:
      consumes:
      - application/json
      description: Get summary of stock reserves
      parameters:
      - description: Reserve Summary Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pb.GetStockReserveSummaryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.ReserveStocksSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Get Stock Reserve Summary
      tags:
      - Reserve
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and then your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"